Add a source for Fano tests under magnetic fields.
Separated the implementation into a source in its own right. It was initially
implemented as an option of the isotropic source. For now it generates particles
isotropically and uses a rejection to generate particles proportional to the mass
of the media so that the Fano test applies in the presence of magnetic fields.
Still need to add the possibility of generating non-isotropically.