Fix several egs_radionuclide_source bugs
Fix a bug that occured when all gammas for a given
excited daughter state had zero probability of emission.

Fix a bug with electron capture where returning a zero
energy particle threw off the particle count. Now
getNextParticle() loops until it gets a particle with
nonzero energy.

Fix some floating point errors.

Fix a bug where all X-Rays had an energy a factor of 10
too large.

Remove use of the branch multiplier for ensdf normalization
records. I'm not sure if this is correct in general but
works for the ensdf files provided by lnhb, for now.

Add a check for existence of ensdf files so that we
crash gracefully.

The 'charge' input is now used correctly.

Improve sampling efficiency of getNextParticle().

Fix a bug in the fluence (it was off by 1).