Fix egs_radionuclide ensdf xray intensities
Fix problem with the energies of x-rays from
radionuclide sources using ensdf files. In the cases
where x-ray data was provided in ensdf comments with a
line denoting the 'total' contribution from a series
of lines, we no longer use the total. Instead,
each line is assigned its own probability. If the
intensities were provided, they are used. If not,
then each line is assigned an equal fraction of the
intensity from the "total" line.

This change avoids a problem where a single x-ray line
(the total) could have an artificially large intensity,
instead of several lower intensity lines. It also avoids
having an x-ray line with a non-physical energy, since
the 'total' lines may be an average of several lines.